restructured files and folders for my liquidity
